feature/crud-courses

+++ routes/api.php
@@ -7,55 +7,55 @@ Route::get('/user', function (Request $request) {
     return $request->user();
 })->middleware('auth:sanctum');
 
-
 Route::prefix('admin')->group(function () {
     Route::prefix('courses')->group(function () {
-        Route::get('/', [\App\Http\Controllers\CourseController::class,'index']);
-        Route::patch('/{id}',[\App\Http\Controllers\CourseController::class,'update']);
-        Route::post('/',[\App\Http\Controllers\CourseController::class,'store']);
-        Route::delete('/',[\App\Http\Controllers\CourseController::class,'destroy']);
+        Route::get('/', [\App\Http\Controllers\Admin\CourseController::class, 'index']);
+        Route::get('/{id}', [\App\Http\Controllers\Admin\CourseController::class, 'show']);
+        Route::patch('/{id}', [\App\Http\Controllers\Admin\CourseController::class, 'update']);
+        Route::post('/', [\App\Http\Controllers\Admin\CourseController::class, 'store']);
+        Route::delete('/', [\App\Http\Controllers\Admin\CourseController::class, 'destroy']);
     });
 
     Route::prefix('users')->group(function () {
-        Route::get('/',[\App\Http\Controllers\UserController::class,'index']);
-        Route::patch('/{id}',[\App\Http\Controllers\UserController::class,'update']);
-        Route::post('/',[\App\Http\Controllers\UserController::class,'store']);
-        Route::delete('/{ids}',[\App\Http\Controllers\UserController::class,'destroy']);
+        Route::get('/', [\App\Http\Controllers\Admin\UserController::class, 'index']);
+        Route::patch('/{id}', [\App\Http\Controllers\Admin\UserController::class, 'update']);
+        Route::post('/', [\App\Http\Controllers\Admin\UserController::class, 'store']);
+        Route::delete('/{ids}', [\App\Http\Controllers\Admin\UserController::class, 'destroy']);
     });
 
     Route::prefix('lessons')->group(function () {
-        Route::get('/',[\App\Http\Controllers\LessonController::class,'index']);
-        Route::patch('/{id}',[\App\Http\Controllers\LessonController::class,'update']);
-        Route::post('/',[\App\Http\Controllers\LessonController::class,'store']);
-        Route::delete('/{ids}',[\App\Http\Controllers\LessonController::class,'destroy']);
+        Route::get('/', [\App\Http\Controllers\Admin\LessonController::class, 'index']);
+        Route::patch('/{id}', [\App\Http\Controllers\Admin\LessonController::class, 'update']);
+        Route::post('/', [\App\Http\Controllers\Admin\LessonController::class, 'store']);
+        Route::delete('/{ids}', [\App\Http\Controllers\Admin\LessonController::class, 'destroy']);
     });
 
     Route::prefix('/academic-years')->group(function () {
-        Route::get('/',[\App\Http\Controllers\AcademicYearController::class,'index']);
-        Route::post('/',[\App\Http\Controllers\AcademicYearController::class,'store']);
-        Route::delete('/{ids}',[\App\Http\Controllers\AcademicYearController::class,'destroy']);
-        Route::patch('/{id}',[\App\Http\Controllers\AcademicYearController::class,'update']);
+        Route::get('/', [\App\Http\Controllers\Admin\AcademicYearController::class, 'index']);
+        Route::post('/', [\App\Http\Controllers\Admin\AcademicYearController::class, 'store']);
+        Route::delete('/{ids}', [\App\Http\Controllers\Admin\AcademicYearController::class, 'destroy']);
+        Route::patch('/{id}', [\App\Http\Controllers\Admin\AcademicYearController::class, 'update']);
     });
 
     Route::prefix('/lesson-tasks')->group(function () {
-        Route::get('/',[\App\Http\Controllers\LessonTaskController::class,'index']);
-        Route::post('/',[\App\Http\Controllers\LessonTaskController::class,'store']);
-        Route::delete('/{ids}',[\App\Http\Controllers\LessonTaskController::class,'destroy']);
-        Route::patch('/{id}',[\App\Http\Controllers\LessonTaskController::class,'update']);
+        Route::get('/', [\App\Http\Controllers\Admin\LessonTaskController::class, 'index']);
+        Route::post('/', [\App\Http\Controllers\Admin\LessonTaskController::class, 'store']);
+        Route::delete('/{ids}', [\App\Http\Controllers\Admin\LessonTaskController::class, 'destroy']);
+        Route::patch('/{id}', [\App\Http\Controllers\Admin\LessonTaskController::class, 'update']);
     });
 
 })->middleware(['auth:sanctum', 'role:'.\App\Models\User::ROLE_ADMIN]);
 
 Route::prefix('departments')->group(function () {
     Route::prefix('courses')->group(function () {
-        Route::get('/',[\App\Http\Controllers\CourseController::class,'index']);
-        Route::get('/{id}',[\App\Http\Controllers\CourseController::class,'show']);
-        Route::post('/',[\App\Http\Controllers\CourseController::class,'store']);
-        Route::patch('/{id}',[\App\Http\Controllers\CourseController::class,'update']);
+        Route::get('/', [\App\Http\Controllers\Admin\CourseController::class, 'index']);
+        Route::get('/{id}', [\App\Http\Controllers\Admin\CourseController::class, 'show']);
+        Route::post('/', [\App\Http\Controllers\Admin\CourseController::class, 'store']);
+        Route::patch('/{id}', [\App\Http\Controllers\Admin\CourseController::class, 'update']);
     });
 });
 
 Route::prefix('auth')->group(function () {
-    Route::post('/login',[\App\Http\Controllers\AuthController::class,'login']);
-    Route::post('/forget-password',[\App\Http\Controllers\AuthController::class,'forget-password']);
+    Route::post('/login', [\App\Http\Controllers\Admin\AuthController::class, 'login']);
+    Route::post('/forget-password', [\App\Http\Controllers\Admin\AuthController::class, 'forget-password']);
 });
